move peer service
