+++ p2p/peer.go
@@ -9,7 +9,7 @@ import (
 	"github.com/it-chain/it-chain-Engine/common"
 	"github.com/it-chain/midgard"
 )
-
+type PeerTable map[string]Peer
 // PeerId 선언
 type PeerId struct {
 	Id string
