--- p2p/peer_table.go
@@ -1,3 +1,0 @@
-package p2p
-
-type PeerTable map[string]Peer
